function/requirements.txt
anthropic>=0.40.0
azure-functions>=1.21.0
requests>=2.32.0
